separated unit tests for card component
